backend/requirements.txt
python-dotenv==0.20.0

Django==4.2.6
mysqlclient==2.2.0

django-extensions==3.2.3

django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

channels==3.0.5
channels_redis==4.2.1
msgpack==1.1.0
redis==5.2.0
async-timeout==5.0.1

PuLP==2.8.0

openpyxl==3.1.2

daphne==3.0.1
